fix bugs for add operation

+++ src/main/java/com/linksteady/operate/vo/LcSpuVO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Data
 public class LcSpuVO {
 
-    private BigDecimal spuWid;
+    private Long spuWid;
 
 
     private String spuName;
